feat: Rename Extracurricular to Eskul and update references
This commit renames `Extracurricular` to `Eskul` across the application.
This includes updating type definitions, component names, and service function calls to reflect the new terminology.

Additionally, mock data for classes has been updated to be an array of objects without IDs, suitable for database seeding. A new feature to seed the database with mock data has also been introduced in the DashboardContent component.